+++ mizani/utils.py
@@ -1,8 +1,23 @@
+from __future__ import annotations
+
+import typing
 from collections import OrderedDict, defaultdict
 from collections.abc import Iterator
 from itertools import chain
+from warnings import warn
+
+try:
+    from zoneinfo import ZoneInfo
+except ImportError:
+    # python < 3.9
+    from backports.zoneinfo import ZoneInfo
 
 import numpy as np
+
+if typing.TYPE_CHECKING:
+    from datetime import datetime
+    from typing import Sequence
+
 
 __all__ = [
     "round_any",
@@ -14,6 +29,7 @@ __all__ = [
     "same_log10_order_of_magnitude",
     "identity",
     "get_categories",
+    "get_timezone",
 ]
 
 DISCRETE_KINDS = "ObUS"
@@ -312,3 +328,46 @@ def log(x, base):
     else:
         res = np.log(x) / np.log(base)
     return res
+
+
+def get_timezone(x: Sequence[datetime]) -> ZoneInfo:
+    """
+    Return a single timezone for the sequence of datetimes
+
+    Returns the timezone of first item and warns if any other item
+    has a different timezone
+    """
+
+    from pytz.tzinfo import BaseTzInfo
+
+    # Ref: https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+
+    info = x[0].tzinfo
+    if info is None:
+        return ZoneInfo("UTC")
+
+    # We recognise two sources of timezone info:
+    #   1. zoneinfo
+    #   2. pytz
+    # These are all subclasses of datetime.tzinfo but they store the
+    # timezone identifier e.g. (Africa/Uganda) in different properties
+    # We assume that all items in the sequence do not mix
+    if isinstance(info, ZoneInfo):
+        prop = "key"
+    elif isinstance(info, BaseTzInfo):
+        prop = "zone"
+    else:
+        raise ValueError(f"Unrecognised timezone class {info.__class__}")
+
+    # Consistency check
+    tz = ZoneInfo(getattr(info, prop))
+    tz_ids = (getattr(value.tzinfo, prop) for value in x)
+    if any(id != tz.key for id in tz_ids):
+        msg = (
+            "Dates in column have different time zones. "
+            "Choosen `{}` the time zone of the first date. "
+            "To use a different time zone, create a "
+            "formatter and pass the time zone."
+        )
+        warn(msg.format(tz.key))
+    return tz